+++ lobster/tools/json/json.py
@@ -21,7 +21,7 @@ import sys
 import json
 from pathlib import PurePath
 from pprint import pprint
-from typing import Tuple, List
+from typing import Tuple, List, Set
 
 from lobster.tool import LOBSTER_Per_File_Tool
 from lobster.items import Tracing_Tag, Activity
@@ -81,23 +81,54 @@ class LOBSTER_Json(LOBSTER_Per_File_Tool):
             description = "Extract tracing data from JSON files.",
             extensions  = ["json"],
             official    = True)
-        self.add_argument("--test-list",
-                          default = "",
-                          help    = ("Member name indicator resulting in a"
-                                     " list containing objects carrying test"
-                                     " data."))
-        self.add_argument("--name-attribute",
-                          default = None,
-                          help    = "Member name indicator for test name.")
-        self.add_argument("--tag-attribute",
-                          default  = None,
-                          required = True,
-                          help     = ("Member name indicator for test "
-                                      " tracing tags."))
-        self.add_argument("--justification-attribute",
-                          default  = None,
-                          help     = ("Member name indicator for "
-                                      " justifications."))
+
+    # Supported config parameters for lobster-json
+    TEST_LIST = "test_list"
+    NAME_ATTRIBUTE = "name_attribute"
+    TAG_ATTRIBUTE = "tag_attribute"
+    JUSTIFICATION_ATTRIBUTE = "justification_attribute"
+    SINGLE = "single"
+
+    @classmethod
+    def get_config_keys_manual(cls):
+        help_dict = super().get_config_keys_manual()
+        help_dict.update(
+            {
+                cls.TEST_LIST: "Member name indicator resulting in a "
+                               "list containing objects carrying test "
+                               "data.",
+                cls.NAME_ATTRIBUTE: "Member name indicator for test name.",
+                cls.TAG_ATTRIBUTE: "Member name indicator for test tracing tags.",
+                cls.JUSTIFICATION_ATTRIBUTE: "Member name indicator for "
+                                             "justifications.",
+                cls.SINGLE: "Avoid use of multiprocessing."
+            }
+        )
+        return help_dict
+
+    def get_mandatory_parameters(self) -> Set[str]:
+        return {self.TAG_ATTRIBUTE}
+
+    def process_commandline_and_yaml_options(
+            self,
+    ) -> Tuple[argparse.Namespace, List[Tuple[File_Reference, str]]]:
+        """
+        Overrides the parent class method and add fetch tool specific options from the
+        yaml
+        config
+
+        Returns
+        -------
+        options - command-line and yaml options
+        worklist - list of json files
+        """
+        options, work_list = super().process_commandline_and_yaml_options()
+        options.test_list = self.config.get(self.TEST_LIST, '')
+        options.name_attribute = self.config.get(self.NAME_ATTRIBUTE)
+        options.tag_attribute = self.config.get(self.TAG_ATTRIBUTE)
+        options.justification_attribute = self.config.get(self.JUSTIFICATION_ATTRIBUTE)
+        options.single = self.config.get(self.SINGLE, False)
+        return options, work_list
 
     def process_tool_options(
             self,